Collapsed leaflet-cartodb-layer-group-view and leaflet-cartodb-group-layer-base
